+++ hspec-core/test/Test/Hspec/Core/Config/OptionsSpec.hs
@@ -53,10 +53,6 @@ spec = do
       it "sets configDiff to False" $ do
         configDiff <$> parseOptions [] Nothing ["--no-diff"] `shouldBe` Right False
 
-    context "with --out" $ do
-      it "sets configOutputFile" $ do
-        either (const Nothing) Just . configOutputFile <$> parseOptions [] Nothing ["--out", "foo"] `shouldBe` Right (Just "foo")
-
     context "with --qc-max-success" $ do
       context "when given an invalid argument" $ do
         it "returns an error message" $ do